test: moved unit tests to 'unit' folder
