+++ tests/unit/commands/context-menus/user-command.decorator.spec.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { ApplicationCommandType } from 'discord-api-types/v10';
 
-import { UserCommand } from '../../../src';
-import { ContextMenuMeta } from '../../../src/commands/context-menus/context-menu.interface';
-import { CONTEXT_MENU_METADATA } from '../../../src/necord.constants';
+import { UserCommand } from '../../../../src';
+import { ContextMenuMeta } from '../../../../src/commands/context-menus/context-menu.interface';
+import { CONTEXT_MENU_METADATA } from '../../../../src/necord.constants';
 
 describe('Context Menus > UserCommand Decorator', () => {
 	const meta: Omit<ContextMenuMeta, 'type'> = {
