+++ tests/unit/commands/slash-commands/slash-command.decorator.spec.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { ApplicationCommandType } from 'discord-api-types/v10';
 
-import { SlashCommand } from '../../../src';
-import { SlashCommandMeta } from '../../../src/commands/slash-commands/slash-command.interface';
-import { SLASH_COMMAND_METADATA } from '../../../src/necord.constants';
+import { SlashCommand } from '../../../../src';
+import { SlashCommandMeta } from '../../../../src/commands/slash-commands/slash-command.interface';
+import { SLASH_COMMAND_METADATA } from '../../../../src/necord.constants';
 
 describe('Slash Commands > SlashCommand Decorator', () => {
 	const meta: SlashCommandMeta = {
