+++ tests/unit/webhooks/necord-webhook-service.decorator.spec.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
-import { NecordWebhookService } from '../../src';
-import { NECORD_WEBHOOK_SERVICE } from '../../src/necord.constants';
+import { NecordWebhookService } from '../../../src';
+import { NECORD_WEBHOOK_SERVICE } from '../../../src/necord.constants';
 
 describe('NecordWebhookService Decorator', () => {
 	@Injectable()
